# build/resources/main/openapi.yaml
openapi: "3.0.3"
...
info:
    title: Beverage Store Service
    version: "1.0"
    description: | # | for multi line description
        Summer is coming and we are all looking forward to having a barbecue with friends.
        For a nice barbecue, we also need some refreshing drinks. Therefore, we want to start a
        new business where we sell beverages and deliver them to people.
    contact:
        name: Group 6
        url: https://gitlab.rz.uni-bamberg.de/dsg/soa/group6
    license:
        name: Apache 2.0
        url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
    - url: http://localhost:9999/v1
      description: Development Server
components:
    schemas:
        BottleDTO:
            type: object
            required:
                - name
                - volume
                - isAlcoholic
                - volumePercent
                - price
                - inStock
            properties:
                id:
                    type: integer
                name:
                    type: string
                volume:
                    type: number
                isAlcoholic:
                    type: boolean
                price:
                    type: number
                inStock:
                    type: integer
                volumePercent:
                    type: number
                supplier:
                    type: string
                href:
                    type: string
...
        Pagination:
            type: object
            properties:
                page:
                    type: number
                noOfPages:
                    type: number
                first:
                    type: string
                previous:
                    type: string
                next:
                    type: string
                last:
                    type: string
        PaginatedBottles:
            type: object
            properties:
                pagination:
                    $ref: "#/components/schemas/Pagination"
                bottles:
                    type: array
                    items:
                        $ref: "#/components/schemas/BottleDTO"
                href:
                    type: string
...
        ErrorMessage:
            type: object
            properties:
                errorType:
                    type: string
                message:
                    type: string

# resources (one of the REST principles)
paths:
    /bottle/list:
        get:
            summary: Get All Bottles
            # request parameters
            parameters:
                - name: page
                  in: query
                  schema:
                      type: number
                      default: 1
                - name: pageLimit
                  in: query
                  schema:
                      type: number
                      default: 10
            # possible responses and their parameters
            responses:
                "200":
                    description: List of all bottles
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PaginatedBottles"
                "400":
                    description: A bad request initiated by the user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
            tags:
                - Bottle Endpoints
    /bottle/search:
        get:
            summary: Bottle Search
            # request parameters
            parameters:
                - name: page
                  in: query
                  schema:
                      type: number
                      default: 1
                - name: pageLimit
                  in: query
                  schema:
                      type: number
                      default: 10
                - name: minPrice
                  in: query
                  schema:
                      type: number
                - name: maxPrice
                  in: query
                  schema:
                      type: number
            responses:
                "200":
                    description: List of all bottles
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PaginatedBottles"
                "400":
                    description: Message body was empty
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
            tags:
                - Bottle Endpoints
    /bottle/{bottleId}:
        get:
            summary: Get a Bottle By ID
            # request parameters
            parameters:
                - name: bottleId
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Bottle Response
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BottleDTO"
                "400":
                    description: Message body was empty
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
            tags:
                - Bottle Endpoints
    /bottle/add:
        post:
            summary: Create New Bottle
            # request parameters
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/BottleDTO"
            responses:
                "200":
                    description: Bottle Response
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BottleDTO"
                "400":
                    description: Message body was empty
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
            tags:
                - Bottle Endpoints
    /bottle/update/{bottleId}:
        put:
            summary: Update Bottle
            # request parameters
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/BottleDTO"
            parameters:
                - name: bottleId
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: Updated Bottle Response
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/BottleDTO"
                "400":
                    description: Message body was empty
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
            tags:
                - Bottle Endpoints
...